Tranfer gentInfo to leaderboard
Set all the fetch responsability outside the scenes
