+++ src/Scenes/CreditsScene.js
@@ -48,7 +48,6 @@ export default class CreditsScene extends Phaser.Scene {
       duration: 8000,
       delay: 1000,
       onComplete: function () {
-        this.madeByTween.destroy;
         this.scene.start('Title');
       }.bind(this),
     });
