+++ src/Scenes/OverScene.js
@@ -1,5 +1,6 @@
 import Phaser from 'phaser';
 import Button from '../Objects/Button';
+import leaderboard from '../leaderboard';
 
 const url = 'https://us-central1-js-capstone-backend.cloudfunctions.net/api/games/SltkTgE89OcOt0kPoJmn/scores/';
 
@@ -19,38 +20,9 @@ export default class OverScene extends Phaser.Scene {
     this.menuButton = new Button(this, 550, 500, 'blueButton1', 'blueButton2', 'Menu', 'Title');
 
 
-    const result = this.getInfo();
+    const result = leaderboard.getInfo();
     result.then((leaderboardDiv) => {
-      const textTest = this.add.dom(400, 220, leaderboardDiv);
+      this.add.dom(400, 220, leaderboardDiv);
     });
   }
-
-  getInfo() {
-    const promiseRefreshScore = fetch(url, {
-      mode: 'cors',
-    });
-    const leaderboardDiv = promiseRefreshScore.then((response) => response.json()).then((response) => {
-      const sorted = response.result.sort((a, b) => {
-        if (a.score < b.score) {
-          return 1;
-        }
-        if (a.score > b.score) {
-          return -1;
-        }
-        return 0;
-      });
-      return sorted;
-    }).then((sorted) => {
-      const arrayTop = sorted.slice(0, 6);
-      const divLeaderboard = document.createElement('div');
-      arrayTop.forEach((score, index) => {
-        const p = document.createElement('p');
-        p.innerText = `#${index + 1} - ${score.user}  ................. ${score.score}`;
-        divLeaderboard.appendChild(p);
-      });
-      return divLeaderboard;
-    });
-    return leaderboardDiv;
-  }
-
 }
