+++ src/leaderboard.js
@@ -20,8 +20,36 @@ const leaderboard = (() => {
     return promiseAddScore;
   };
 
+  const getInfo = () => {
+    const promiseRefreshScore = fetch(url, {
+      mode: 'cors',
+    });
+    const leaderboardDiv = promiseRefreshScore.then((response) => response.json()).then((response) => {
+      const sorted = response.result.sort((a, b) => {
+        if (a.score < b.score) {
+          return 1;
+        }
+        if (a.score > b.score) {
+          return -1;
+        }
+        return 0;
+      });
+      return sorted;
+    }).then((sorted) => {
+      const arrayTop = sorted.slice(0, 6);
+      const divLeaderboard = document.createElement('div');
+      arrayTop.forEach((score, index) => {
+        const p = document.createElement('p');
+        p.innerText = `#${index + 1} - ${score.user}  ................. ${score.score}`;
+        divLeaderboard.appendChild(p);
+      });
+      return divLeaderboard;
+    });
+    return leaderboardDiv;
+  };
 
-  return { addScore };
+
+  return { addScore, getInfo };
 })();
 
 export default leaderboard;
